% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkundidversion.R
\name{checkundidversion}
\alias{checkundidversion}
\title{Checks and prints the currently installed and the latest version of Undid.jl package}
\usage{
checkundidversion()
}
\description{
Checks and prints the currently installed and the latest version of Undid.jl package
}
