% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_diff_df.R
\name{create_diff_df}
\alias{create_diff_df}
\title{Creates the empty_diff_df.csv to be sent out to all the silos.}
\usage{
create_diff_df(
  filepath,
  date_format,
  freq,
  covariates = FALSE,
  freq_multiplier = FALSE
)
}
\arguments{
\item{filepath}{A string filepath to the init.csv}

\item{date_format}{A string indicating the format of the dates in the init.csv}

\item{freq}{A string indicating the frequency of the data to be analyzed (e.g. monthly, yearly)}

\item{covariates}{A character vector of covariates (optional) or \code{FALSE} to exclude to use covariates found in the init.csv}

\item{freq_multiplier}{An integer, if required, to multiply with the freq (e.g. if you are analyzing data that is collected every 3 months you would set freq_multiplier = 3 and freq = "monthly")}
}
\description{
The \code{create_diff_df} function generates a CSV file with missing values for that are then filled at the appropriate silos.
}
\details{
Ensure dates are entered consistently in the same format.
Control silos should be marked as "control" in the \code{treatment_times} column.
}
\examples{
df = create_diff_df("C:/Users/User/Project Files/init.csv", "yyyy", "yearly")
}
