% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_init_csv.R
\name{create_init_csv}
\alias{create_init_csv}
\title{Creates an initial .csv file (init.csv) specifying the silos, start times, end times, and treatment (or lack thereof) times.}
\usage{
create_init_csv(
  silo_names = NA,
  start_times = NA,
  end_times = NA,
  treatment_times = NA,
  covariates = FALSE
)
}
\arguments{
\item{silo_names}{A character vector of silo names (e.g., "71", "73").}

\item{start_times}{A character vector of start times, corresponding to the silos.}

\item{end_times}{A character vector of end times, corresponding to the silos.}

\item{treatment_times}{A character vector of treatment times or "control", indicating when treatment started for each silo.}

\item{covariates}{A character vector of covariates (optional) or \code{FALSE} to exclude covariates.}
}
\description{
The \code{create_init_csv} function generates a CSV file with silo data, including start times, end times, and treatment times.
The function can generate a blank CSV with just the headers, or it can take specific data inputs to pre-fill the CSV.
}
\details{
Ensure dates are entered consistently in the same format.
Control silos should be marked as "control" in the \code{treatment_times} column.
}
\examples{
# Example of how to use the create_init_csv function:
silo_names <- c("71", "73")
start_times <- c("1989", "1989")
end_times <- c("2000", "2000")
treatment_times <- c("1991", "control")
covariates <- c("asian", "black", "male")

# Generate and save the CSV file, and return the data frame
df <- create_init_csv(silo_names, start_times, end_times, treatment_times, covariates)

# Output the resulting data frame
print(df)
}
