% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_parallel_trends.R
\name{plot_parallel_trends}
\alias{plot_parallel_trends}
\title{Plots parallel trends figures.}
\usage{
plot_parallel_trends(
  dir_path,
  covariates = FALSE,
  save_csv = FALSE,
  combine = FALSE,
  pch = NA,
  pch_control = NULL,
  pch_treated = NULL,
  control_colour = c("darkgrey", "lightgrey"),
  control_color = NULL,
  treatment_colour = c("darkred", "lightcoral"),
  treatment_color = NULL,
  lwd = 2,
  xlab = NA,
  ylab = NA,
  title = NA,
  xticks = 4,
  date_format = "\%Y",
  xdates = NULL,
  xaxlabsz = 0.8,
  save_png = FALSE,
  width = 800,
  height = 600,
  ylim = NULL,
  yaxlabsz = 0.8,
  ylabels = NULL,
  yticks = 4,
  ydecimal = 2,
  legend_location = "topright",
  simplify_legend = TRUE,
  legend_text = 0.7,
  legend_on = TRUE,
  treatment_indicator_col = "grey",
  treatment_indicator_alpha = 0.5,
  treatment_indicator_lwd = 2,
  treatment_indicator_lty = 2
)
}
\arguments{
\item{dir_path}{A string filepath to folder containing all of the trends_data.csv's}

\item{covariates}{A logical value (defaults to \code{FALSE}) indicating whether or not to consider covariates}

\item{save_csv}{A logical value (defaults to \code{FALSE}) indicating whether or not to save the combined_trends_data.csv}

\item{combine}{A logical value (defaults to \code{FALSE}) indicating whether to plot each silo separately or to combine silos based on treatment status}

\item{pch}{An integer (from 0 to 25, defaults to \code{NA}) or vector of integers (from 0 to 25) determines the points used on the plot}

\item{pch_control}{An integer (from 0 to 25, defaults to \code{NULL}) or vector of integers (from 0 to 25) determines the points used on the plot for control silos. Takes value of pch if NULL}

\item{pch_treated}{An integer (from 0 to 25, defaults to \code{NULL}) or vector of integers (from 0 to 25) determines the points used on the plot for treated silos. Takes value of pch if NULL}

\item{control_colour}{A vector of string colours (defaults to \code{c("darkgrey", "lightgrey")}) option for the control silos. If combine = TRUE, takes the 1st value to determine colour of control line}

\item{treatment_colour}{A vector of string colours (defaults to \code{c("darkred", "lightcoral")}) for the treatment silos. If combine = TRUE, takes the 1st value to determine colour of control line}

\item{lwd}{An integer (defaults to \code{2}) for selecting the line widths}

\item{xlab}{A string value for the x-axis label (defaults to \code{NA})}

\item{ylab}{A string value for the y-axis label (defaults to \code{NA})}

\item{title}{A string value for the title of the plot (defaults to \code{NA})}

\item{xticks}{An integer value denoting how many ticks to display on the x-axis (defaults to \code{4})}

\item{date_format}{A string value denoting the format with which to display the dates along the x-axis (defaults to \code{"\%Y"})}

\item{xdates}{Takes in a vector of dates to be used at the dates shown along the x-axis (defaults to \code{NULL})}

\item{xaxlabsz}{A double indicating the x-axis label sizes (defaults to \code{0.8})}

\item{save_png}{A logical value indicating whether or not to save the plot as a png file (defaults to \code{FALSE})}

\item{width}{An integer denoting the width of the saved .png file}

\item{height}{An integer denoting the height of the saved .png file}

\item{ylim}{A vector of two doubles defining the min and max range of the values on the y-axis. Defaults to the min and max values in the y column of trends_data}

\item{yaxlabsz}{A double for specifying the y-axis label sizes (defaults to \code{0.8})}

\item{ylabels}{A vector of values that you would like to appear on the y-axis (defaults to \code{NULL})}

\item{yticks}{An integer denoting how many values to display along the y-axis (defaults to \code{4})}

\item{ydecimal}{An integer value denoting to which decimal point the values along the y-axis are rounded to}

\item{legend_location}{A string value for determining the location of the legend (defaults to \code{topright})}

\item{simplify_legend}{A logical value which if set to \code{TRUE} shows one colour for the treatment silos in the legend and one colour for the control. Defaults to \code{TRUE}}

\item{legend_text}{An double for setting the size of the text in the legend. Defaults to \code{0.7}}

\item{legend_on}{A logical value for turning the legend on or off (defaults to \code{TRUE})}

\item{treatment_indicator_col}{A string value for determining the colour of the dashed vertical line showing when treatment times were (defaults to \code{"grey"})}

\item{treatment_indicator_alpha}{A double for for determining the transparency level of the dashed vertical lines showing the treatment times (defaults to \code{0.5})}

\item{treatment_indicator_lwd}{A double for for selecting the line width of the treatment indicator lines (defaults to \code{2})}

\item{treatment_indicator_lty}{An integer for the selecting the lty option for the treatment_indicator lines (defaults to \code{2})}
}
\description{
The \code{plot_parallel_trends} function combines the trends_data.csv's and plots parallel trends figures.
}
\details{
All treatment and all control groups can be combined so that there is one control line and one treatment line by setting combine = \code{TRUE}
}
\examples{
trends_data <- plot_parallel_trends("C:/Users/User/Downloads/")

}
