% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undid_stage_three.R
\name{undid_stage_three}
\alias{undid_stage_three}
\title{Compute the aggregate ATT.}
\usage{
undid_stage_three(
  dir_path,
  agg = "silo",
  covariates = FALSE,
  save_csv = FALSE,
  interpolation = FALSE
)
}
\arguments{
\item{dir_path}{A string filepath to folder containing all of the filled_diff_df.csv's}

\item{agg}{A string specifying the aggregation method. Either "silo", "g", or "gt".}

\item{covariates}{A logical value (default \code{FALSE}), indicating whether to consider covariates}

\item{save_csv}{A logical value (default \code{FALSE}), indicating whether to save the combined_diff_df.csv}

\item{interpolation}{Takes in either \code{FALSE} or "linear_function" as options. Defaults to \code{FALSE}. Used to fill in missing values in the filled_diff_df.csv's}
}
\description{
The \code{undid_stage_three} function combines the filled_diff_df.csv's and computes the aggregate ATT by silo, gvar, or (g,t).
}
\details{
If errors persist, check to see if there are missing diff_estiamtes in the filled_diff_df.csv's. If so, try setting interpolation = "linear_function"
}
\examples{
output <- undid_stage_three(filepath = "C:/Users/User/Downloads/", 
interpolation = "linear_function")

# To view the df
print(output)

}
