% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undid_stage_two.R
\name{undid_stage_two}
\alias{undid_stage_two}
\title{Fill in the empty_diff_df.csv with the relevant diff_estimate values at each silo.}
\usage{
undid_stage_two(
  filepath,
  local_silo_name,
  silo_data,
  time_column,
  outcome_column,
  local_date_format,
  consider_covariates = TRUE
)
}
\arguments{
\item{filepath}{A string filepath to the empty_diff_df.csv}

\item{local_silo_name}{A string specifying the name of the local silo as it is written in the empty_diff_df.csv}

\item{silo_data}{A dataframe containing the local silo's data}

\item{time_column}{A string specifying the name of the column with the time/date information}

\item{outcome_column}{A string specifying the name of the column with outcome of interest data}

\item{local_date_format}{A string indicating the format of the dates in the data}

\item{consider_covariates}{A logical value (default \code{TRUE}), indicating whether to consider covariates}
}
\description{
The \code{undid_stage_two} function uses information on dates from the empty_diff_df.csv and the local silo data to fill in the necessary diff_estimates.
}
\details{
details here
}
\examples{
silo_data <- read_dta("C:/Users/User/Documents/Data/State71.dta")
output <- undid_stage_two(filepath = "C:/Users/User/Downloads/empty_diff_df.csv", 
local_silo_name = "71", 
silo_data = silo_data, 
time_column = "year", 
outcome_column = "coll", 
local_date_format = "yyyy")

# To view the filled_diff_data:
print(output[[1]][2])

# To view the trends_data
print(output[[2]][2])

}
