% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateundid.R
\name{updateundid}
\alias{updateundid}
\title{Updates Undid.jl to the latest version if Undid.jl is already installed.}
\usage{
updateundid()
}
\description{
Updates Undid.jl to the latest version if Undid.jl is already installed.
}
